+++ versions.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "~> 3.23"
+      version = "~> 3.28"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"